Move BootLoaderImage to pyanaconda.bootloader.image
Move the class BootLoaderImage and the related code to a new file.
